reorganize by taking alteration code out of base
